FIx to handle source files in nested directories
Modification to handle source files which reside in directories nested within a primary build directory and are to be considered part of the primary build directory target make. This was needed to support a directory strategy Stephen Taylor used for the Stream Financial source code which was based upon the 'folder' approach taken in Microsoft in Visual Studio.